refactor(bootstrap): Move runtime bridge defaults into bootstrap
Move the default production bridge binding selection out of the runtime holder and endpoint helper packages into a bootstrap-owned helper. Remove the obsolete endpoint-owned binding helpers, update bootstrap call sites and package docs, and move the affected coverage to the new bootstrap helper test.
