+++ src/main/java/network/crypta/clients/http/bridge/package-info.java
@@ -10,9 +10,10 @@
  *
  * <p>In practical terms, this package owns the default bridge objects that wire {@code
  * SimpleToadletServer}, FProxy support, and shell startup helpers to the daemon core. Production
- * binding selection still starts in {@code network.crypta.runtime.endpoints.http}, but that
- * runtime-owned package should treat these classes as adapter implementations, not as part of the
- * long-lived runtime seam. Keeping the concrete bridge here preserves current behavior while making
- * ownership explicit for later extraction work.
+ * binding selection now starts in {@code
+ * network.crypta.runtime.bootstrap.DefaultNodeRuntimeBridgeFactories}, while runtime-owned packages
+ * continue to treat these classes as adapter implementations rather than as part of the long-lived
+ * runtime seam. Keeping the concrete bridge here preserves current behavior while making ownership
+ * explicit for later extraction work.
  */
 package network.crypta.clients.http.bridge;
