+++ src/main/java/network/crypta/runtime/endpoints/fcp/package-info.java
@@ -1,11 +1,10 @@
 /**
- * Runtime-owned FCP seam bindings and endpoint handle surface.
+ * Runtime-owned FCP endpoint handle surface.
  *
- * <p>This package now contains only the narrow runtime-owned FCP endpoint seam, plus the default
- * production binding entry point that selects the adapter-owned bridge implementations in {@code
- * network.crypta.clients.fcp.bridge}. Higher-level runtime/bootstrap code should depend on {@link
- * network.crypta.runtime.endpoints.fcp.FcpEndpointHandle} and the factory methods on {@link
- * network.crypta.runtime.endpoints.fcp.FcpEndpointBridgeFactories}, not on the concrete bridge
- * classes themselves.
+ * <p>This package now contains only the narrow runtime-owned {@link
+ * network.crypta.runtime.endpoints.fcp.FcpEndpointHandle} seam. Bootstrap-owned code selects the
+ * adapter-backed persistent-request services elsewhere, and the concrete FCP bridge classes remain
+ * in {@code network.crypta.clients.fcp.bridge}. Higher-level runtime code should therefore depend
+ * on this endpoint-handle seam rather than on concrete FCP server types.
  */
 package network.crypta.runtime.endpoints.fcp;
